+++ packages/create-graph/src/index.ts
@@ -1,16 +1,14 @@
-import { GraphActions } from '@magic/graph-primitives/actions/types';
-import { EventHub } from '@magic/graph-primitives/events/createEventHub';
-import { GraphGetters } from '@magic/graph-primitives/getters/types';
-import { CoreEdge, CoreNode } from '@magic/graph-primitives/types';
+import { core as createCore } from '@magic/graph-core/index';
+import { GraphSettings } from '@magic/graph-core/settings/index';
+import { CoreControls } from '@magic/graph-core/types';
 import { createComputedTokenResolver } from '@magic/graph-plugins-shared/computed-tokens/createComputedTokenResolver';
 import {
   ExtractActions,
   ExtractControls,
   ExtractEventMap,
   ExtractGetters,
-  LooseGraphPlugin,
-} from '@magic/graph-plugins-shared/plugins/types';
-import { ThemeController } from '@magic/graph-plugins-shared/theme/createThemeController';
+} from '@magic/graph-plugins-shared/plugins/extractors';
+import { LooseGraphPlugin } from '@magic/graph-plugins-shared/plugins/loose';
 import {
   PluginThemeField,
   ThemesForPlugins,
@@ -21,9 +19,10 @@ import {
 } from '@magic/graph-plugins/canvas/aggregator/types';
 import { CANVAS_ELEMENT_CURSOR_FIELD_KEY } from '@magic/graph-plugins/canvas/setupCanvasCursor';
 import { CanvasControls } from '@magic/graph-plugins/canvas/types';
-import { CoreControls } from '@magic/graph-core/types';
-import { core as createCore } from '@magic/graph-core/index';
-import { GraphSettings } from '@magic/graph-core/settings/index';
+import { GraphActions } from '@magic/graph-primitives/actions/types';
+import { EventHub } from '@magic/graph-primitives/events/createEventHub';
+import { GraphGetters } from '@magic/graph-primitives/getters/types';
+import { CoreEdge, CoreNode } from '@magic/graph-primitives/types';
 import { nullThrows } from '@magic/utils/assert';
 import type { Prettify } from 'ts-essentials';
